refactor/ get_current_weather_data
